Remove old non channel version of TPS2012 driver
